requirements.txt
python-telegram-bot==21.6
requests==2.32.3
urllib3==2.2.2